+++ src/main/java/com/nchowf/tutorlinking/classes/ClassRepo.java
@@ -2,5 +2,5 @@ package com.nchowf.tutorlinking.classes;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClassRepo extends JpaRepository<Class,Long> {
+public interface ClassRepo extends JpaRepository<Class,Integer> {
 }
